+++ packages/wallet-cli/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @lightsparkdev/wallet-cli
 
+## 0.0.19
+
+### Patch Changes
+
+- 6ea8639: Some improvements to the wallet CLI
+
+  - Support JWT signing key generation as part of init-env
+  - Allow safe backup of your existing .lightspark.env if you are initializing a new environment
+  - Add a terminate-wallet command
+
 ## 0.0.18
 
 ### Patch Changes